Exchange: Sync the calendar even if the sync key is null.
For the calendar sync mail box will be saved in the Email, and
controled by the email account sync frequency. Then if the user
set the email sync frequency as never, the calendar sync key is
null. As the default, the calendar will be never synced. But for
user only set the email sync frequency, it is not a good behavior.

So delete the judgement for the value of sync key. And this is
samed as the contact sync function.

CRs-Fixed: 367604

Change-Id: I6be742c7ac4b2c47095ef8f10936c42e14c7e39c